+++ extensions/hashnode/src/components/StoryItem.tsx
@@ -1,12 +1,14 @@
-import { ActionPanel, Icon, List, Action, Image } from "@raycast/api";
+import { ActionPanel, Icon, List, Action, Image, Keyboard } from "@raycast/api";
 import { HashnodeColors } from "../models/Colors";
 import { Story } from "../models/Story";
+import { useCachedState } from "@raycast/utils";
 
 interface StoryItemProps {
   story: Story;
 }
 
 export default function StoryItem({ story }: StoryItemProps) {
+  const [isShowingDetail] = useCachedState("is-showing-detail", false);
   return (
     <List.Item
       icon={{
@@ -16,17 +18,36 @@ export default function StoryItem({ story }: StoryItemProps) {
       }}
       title={story.title ?? "No title"}
       actions={<Actions item={story} />}
-      accessories={[
-        { date: new Date(story.publishedAt) },
-        {
-          text: `👍  ${story.reactionCount}`,
-        },
-      ]}
+      accessories={
+        isShowingDetail
+          ? undefined
+          : [
+              { date: new Date(story.publishedAt) },
+              {
+                text: `👍  ${story.reactionCount}`,
+              },
+            ]
+      }
+      detail={
+        <List.Item.Detail
+          markdown={story.brief}
+          metadata={
+            <List.Item.Detail.Metadata>
+              <List.Item.Detail.Metadata.Link
+                title="Publication"
+                text={story.publication.title}
+                target={story.publication.url}
+              />
+            </List.Item.Detail.Metadata>
+          }
+        />
+      }
     />
   );
 }
 
 function Actions({ item }: { item: Story }) {
+  const [, setIsShowingDetail] = useCachedState("is-showing-detail", false);
   if (item.url) {
     return (
       <ActionPanel title={item.title ?? "No title"}>
@@ -42,6 +63,12 @@ function Actions({ item }: { item: Story }) {
             url={item.publication.url}
           />
         </ActionPanel.Section>
+        <Action
+          icon={Icon.AppWindowSidebarLeft}
+          title="Toggle Details"
+          onAction={() => setIsShowingDetail((prev) => !prev)}
+          shortcut={Keyboard.Shortcut.Common.ToggleQuickLook}
+        />
       </ActionPanel>
     );
   }