+++ extensions/hashnode/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Hashnode Changelog
 
+## [Pagination 🚀 + More] - {PR_MERGE_DATE}
+
+- Empty View if no Articles Published
+- Pagination allows you to doom scroll
+- Toggle a `Details` View with Post Brief and Publication Name
+
 ## [Migrate API + Cache + Modernize] - 2024-09-03
 
 - fix: extension not working as it was using deprecated API